Add files via upload
This python script will get file names from a specific directory, then create an old filenames csv. You will already have to have created a new filenames csv in the same directory. The script with then combine the old filenames with the new filenames csv and change the names of all the files in the directory according to the names you had in the new files csv.